+++ packages/react-dogfood/helpers/logger.ts
@@ -1,7 +1,8 @@
-import { LogLevel, Logger, logToConsole } from '@stream-io/video-react-sdk';
+import { Logger, LogLevel, logToConsole } from '@stream-io/video-react-sdk';
 import * as Sentry from '@sentry/nextjs';
 
 const logLevelMapping = new Map<LogLevel, Sentry.SeverityLevel>();
+logLevelMapping.set('trace', 'debug');
 logLevelMapping.set('debug', 'debug');
 logLevelMapping.set('info', 'info');
 logLevelMapping.set('warn', 'warning');
@@ -10,16 +11,14 @@ logLevelMapping.set('error', 'error');
 export const customSentryLogger: Logger = (
   logLevel: LogLevel,
   message: string,
-  extraData?: any,
-  tags?: string[],
+  ...args: unknown[]
 ) => {
   if (logLevel === 'warn' || logLevel === 'error') {
     Sentry.captureEvent({
       level: logLevelMapping.get(logLevel),
-      extra: extraData,
     });
   }
 
   // Call the SDK's default log method
-  logToConsole(logLevel, message, extraData, tags);
+  logToConsole(logLevel, message, ...args);
 };